test/e2e/requirements.txt
Appium-Python-Client>=3.0
pytest
selenium
colorama
pytest-variables
pytest-selenium
pyyaml